+++ frontend/csi/plugin.go
@@ -5,6 +5,7 @@ package csi
 import (
 	"os"
 	"strings"
+	"sync"
 
 	"github.com/container-storage-interface/spec/lib/go/csi"
 	log "github.com/sirupsen/logrus"
@@ -34,6 +35,8 @@ type Plugin struct {
 	restClient *RestClient
 	helper     helpers.HybridPlugin
 
+	mutex sync.Mutex
+
 	grpc NonBlockingGRPCServer
 
 	csCap []*csi.ControllerServiceCapability
@@ -55,6 +58,7 @@ func NewControllerPlugin(
 		endpoint:     endpoint,
 		role:         CSIController,
 		helper:       *helper,
+		mutex:        sync.Mutex{},
 		opCache:      make(map[string]bool),
 	}